Updates
Util: Fixed saving of upcomming game, and the check if the playerfile
is empty.
MenuSystem: Fixed minor error where we could not create a Keeper, do to
a check where we used “OR” instead of “AND”: || instead of &&
Game: Added an Arraylist<String> for handling the goals of each game,
so that tomorrow or later to day, we can finish the creation of a
PlayedGame. Tomorrow as in dec 4.